#94 - Still need to Bug Fix everything

+++ src/main/resources/db/migration/V1_29052021_0550__Create_User_Table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users
 (
-    `id`            INT         NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    `id`            BIGINT(11) NOT NULL PRIMARY KEY AUTO_INCREMENT,
     `username`      VARCHAR(50) NOT NULL,
     `full_name`     VARCHAR(75) NOT NULL,
     `password_hash` VARCHAR(32) NOT NULL
